refactor(memory): inline _flatten_reasons, drop _errors module
Move the _flatten_reasons helper into memory_manager.py (where the TS source defines it) and delete memory/_errors.py, which had no TypeScript counterpart. AggregateMemoryError remains in strands.types.exceptions.
